requirements.txt
openai>=1.30.0
faiss-cpu
numpy
tqdm
python-dotenv
flask
flask-cors
gunicorn
